Implement Search for FoodTruckCompany

+++ FoodTruckServices/Model/GenericModels/AreaCoordination.cs
@@ -8,7 +8,7 @@ namespace FoodTruckServices.Model
     public class AreaCoordination
     {
         public Coordination Coordination { get; set; }
-        public int Radius { get; set; }
+        public float Radius { get; set; }
 
         public DistanceUnit Unit { get; set; }
     }
